Python_day_06/src/requirements.txt
annotated-types==0.5.0
beautifulsoup4==4.12.2
google==3.0.0
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.3
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
soupsieve==2.5
SQLAlchemy==2.0.21
typing_extensions==4.8.0
